sched: hmp: Assigned bigger value to min_runnable_load
If caller of hmp_domain_min_load does not use return value about min_usage,
per_cpu variable can access area of non-existed cpu(NR_CPUS) and occurs kernel
panic. By assigning bigger value(ULONG_MAX) to min_runnable_load, if some cpu
exists in domain, then hmp_domain_min_load() does not return NR_CPUS.

+++ kernel/sched/fair.c
@@ -5525,7 +5525,7 @@ static inline unsigned int hmp_domain_min_load(struct hmp_domain *hmpd,
 	int min_cpu_runnable_temp = NR_CPUS;
 	u64 min_target_last_migration = ULLONG_MAX;
 	u64 curr_last_migration;
-	unsigned long min_runnable_load = INT_MAX;
+	unsigned long min_runnable_load = ULONG_MAX;
 	unsigned long contrib;
 	struct sched_avg *avg;
 	struct cpumask temp_cpumask;
